Added a dashboard, confirm delete and froup by date

+++ src/tasks/views.py
@@ -3,6 +3,9 @@ from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
+from itertools import groupby
+from django.utils.timezone import localtime
+from django.db.models.functions import TruncDate
 
 from tasks.models import Task
 
@@ -28,7 +31,7 @@ def login_view(request):
         user = authenticate(request, username=username, password=password)
         if user is not None:
             login(request, user)
-            return redirect('task_list')
+            return redirect('dashboard')
         else:
             messages.error(request, 'Invalid credentials')
     return render(request, 'tasks/login.html')
@@ -43,19 +46,86 @@ def task_list(request):
     if request.method == 'POST':
         title = request.POST.get('title')
         if title:
-            Task.objects.create(title=title)
+            Task.objects.create(title=title, user=request.user)
             return redirect('/')
     tasks = Task.objects.all()
     return render(request, 'tasks/task_list.html', {'tasks': tasks})
 
 
-def complete_task(request, pk):
-    task = get_object_or_404(Task, pk=pk)
+def complete_task(request, task_id):
+    task = get_object_or_404(Task, id=task_id)
     task.completed = not task.completed
     task.save()
     return redirect('/')
 
-def delete_task(request, pk):
-    task = get_object_or_404(Task, pk=pk)
-    task.delete()
-    return redirect('/')
+@login_required(login_url='login')
+def delete_task(request, task_id):
+    task = get_object_or_404(Task, id=task_id)
+
+    if request.method == 'POST':
+        task.delete()
+        return redirect('dashboard')
+    
+    return render(request, 'tasks/confirm_delete.html', {'task': task})
+
+
+
+
+@login_required(login_url='login')
+def dashboard(request):
+    tasks = (
+        Task.objects
+        .annotate(created_date=TruncDate('created_at'))
+        .order_by('-created_date')
+    )
+
+    grouped_tasks = {}
+    for date, items in groupby(tasks, key=lambda x: x.created_date): # type: ignore
+        grouped_tasks[date] = list(items)
+
+    total = tasks.count()
+    completed = tasks.filter(completed=True).count()
+    pending = total - completed
+
+    return render(request, 'tasks/dashboard.html', {
+        'grouped_tasks': grouped_tasks,
+        'total': total,
+        'completed': completed,
+        'pending': pending,
+    })
+
+
+   # total = tasks.count()
+    #completed = tasks.filter(completed=True).count()
+   # pending = total - completed
+   # recent_tasks = tasks.order_by('id')[:5]
+
+   # context = {
+    #    'user': user,
+   #     'total': total,
+   #     'completed': completed,
+   #     'pending': pending,
+    #    'recent_tasks': recent_tasks,
+   # }
+
+    #return render(request, 'tasks/dashboard.html', {'grouped_tasks': grouped_tasks})
+
+def edit_task(request, task_id):
+    task = get_object_or_404(Task, id=task_id)
+
+    if request.method == 'POST':
+        title = request.POST.get('title')
+        completed = request.POST.get('completed') == 'on'
+        if title:
+            task.title = title
+            task.completed = completed
+            task.save()
+            return redirect('dashboard')
+        
+    return render(request, 'tasks/edit_task.html', {'task': task})
+
+def toggle_complete(request, task_id):
+     task = get_object_or_404(Task, id=task_id)
+     task.completed = not task.completed
+     task.save()
+     return redirect('dashboard')
